Moving to a parse.go
